+++ CHANGELOG.md
@@ -1,9 +1,12 @@
 # 0.2.0 (2019-04-xx)
-* [WIP] Added mongodb like state update support
+
+* Added mongodb like state update support (via immutability-helper)
+* Changed to be able to set actions to props of Nanox container
 * Changed library format to UMD
 * Update dependencies
 
 # 0.1.3 (2019-03-19)
+
 * Update dependencies
 
 # 0.1.2 (2017-12-27)